ai-platform/services/agent_service/requirements.txt
google-generativeai==0.5.4
pydantic==2.7.4
confluent-kafka==2.4.0
redis==5.0.6
httpx==0.27.0
